trabalho-final/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.4
click==8.1.7
cryptography==3.4.8
distro==1.7.0
exceptiongroup==1.2.2
fastapi==0.115.4
fpdf2==2.8.1
h11==0.14.0
httplib2==0.20.2
idna==3.10
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
pika==1.3.2
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.3.0
pyparsing==2.4.7
six==1.16.0
sniffio==1.3.1
uvicorn==0.32.0
bcrypt==4.2.0
passlib==1.7.4
typing-extensions==4.12.2
starlette==0.41.2
greenlet==3.1.1
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-core==2.23.4
sqlalchemy==2.0.36
passlib==1.7.4
python-multipart==0.0.16
ecdsa==0.19.0
pyasn1==0.6.1
python-jose==3.3.0
rsa==4.9